fix: implement useVerification hook and integrate it into VerificationForm
