+++ src/pages/payrollAgreement/tabs/forms/verificationPayrollAgreement/index.tsx
@@ -1,17 +1,17 @@
 import { MdOutlineArrowBack } from "react-icons/md";
-import { Button, Stack, useMediaQuery } from "@inubekit/inubekit";
+import { Button, Stack } from "@inubekit/inubekit";
 
 import { ComponentAppearance } from "@enum/appearances";
 import { Accordion } from "@design/data/accordions";
 import { tokens } from "@design/tokens";
 import { DecisionModal } from "@design/modals/decisionModal";
-import { addPayrollAgreementSteps } from "@config/payrollAgreement/payrollAgreementTab/assisted/steps";
 import { RequestProcess } from "@design/feedback/RequestProcess";
 import { RequestStatusModal } from "@design/modals/requestStatusModal";
 import { requestProcessMessage } from "@config/payrollAgreement/payrollAgreementTab/generic/requestProcessMessage";
 import { finishModal } from "@config/payrollAgreement/payrollAgreementTab/generic/finishModal";
 import { requestStatusMessage } from "@config/payrollAgreement/payrollAgreementTab/generic/requestStatusMessage";
 import { IVerificationForm } from "@ptypes/payrollAgreement/payrollAgreementTab/forms/IVerificationForm";
+import { useVerification } from "@hooks/payrollAgreement/useVerificationForm";
 import { verificationFormLabels } from "@config/payrollAgreement/payrollAgreementTab/assisted/verificationFormLabels";
 import { VerificationBoxes } from "./verificationBoxes";
 
@@ -33,28 +33,18 @@ const VerificationForm = (props: IVerificationForm) => {
     onClosePendingReqModal,
   } = props;
 
-  const isTablet = useMediaQuery("(max-width: 1224px)");
-
-  const isMobile = useMediaQuery("(max-width: 990px)");
-
-  const canShowRequestProcess = showRequestProcessModal && savePayrollAgreement;
-
-  const canShowPendingRequest =
-    showPendingReqModal &&
-    savePayrollAgreement &&
-    savePayrollAgreement.requestNumber.length > 0;
-
-  const steps = addPayrollAgreementSteps.filter((step) => {
-    if (step.name.toLowerCase() === "verificación") return false;
-    if (
-      Number(step.id) === 4 &&
-      (!updatedData.extraordinaryCycles.values ||
-        updatedData.extraordinaryCycles.values.length === 0)
-    ) {
-      return false;
-    }
-    if (typeRegularPayroll) return true;
-    return Number(step.id) !== 3;
+  const {
+    steps,
+    isTablet,
+    isMobile,
+    canShowRequestProcess,
+    canShowPendingRequest,
+  } = useVerification({
+    updatedData,
+    showRequestProcessModal,
+    savePayrollAgreement,
+    typeRegularPayroll,
+    showPendingReqModal,
   });
 
   return (
